+++ cornac/models/lightgcn/__init__.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 # ============================================================================
 
-from .recom_lightgcn import LightGCN1
+from .recom_lightgcn import LightGCN
