--- cornac/models/lightgcn2/requirements.txt
@@ -1,2 +1,0 @@
-torch>=2.0.0
-dgl>=1.1.0
